# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MetricFlow API",
    "version": "1.0.0",
    "description": "API documentation for MetricFlow Backend"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    },
    {
      "url": "/api",
      "description": "Relative path (Production)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/fees": {
      "get": {
        "summary": "List all fee configurations",
        "tags": [
          "Admin Fees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of fees"
          }
        }
      },
      "post": {
        "summary": "Create a fee configuration",
        "tags": [
          "Admin Fees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "fee_type",
                  "config_type",
                  "config"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "fee_type": {
                    "type": "string",
                    "enum": [
                      "funding_card",
                      "funding_account",
                      "transfer",
                      "otp",
                      "stamp_duty"
                    ]
                  },
                  "config_type": {
                    "type": "string",
                    "enum": [
                      "percentage_cap",
                      "flat",
                      "flat_conditional",
                      "range"
                    ]
                  },
                  "config": {
                    "type": "object"
                  },
                  "currency": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fee created"
          }
        }
      }
    },
    "/admin/fees/{id}": {
      "put": {
        "summary": "Update a fee configuration",
        "tags": [
          "Admin Fees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "fee_type": {
                    "type": "string"
                  },
                  "config_type": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object"
                  },
                  "currency": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fee updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a fee configuration",
        "tags": [
          "Admin Fees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fee deleted"
          }
        }
      }
    },
...
    "/fees": {
      "get": {
        "summary": "View applicable fees",
        "tags": [
          "Fees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of fees"
          }
        }
      }
    },
...
  },
...
}